exception handling added

+++ src/main/java/com/asadabdalla/soccerplayer/services/PlayerService.java
@@ -1,5 +1,7 @@
 package com.asadabdalla.soccerplayer.services;
 
+import com.asadabdalla.soccerplayer.exceptions.BadRequestException;
+import com.asadabdalla.soccerplayer.exceptions.PlayerNotFoundException;
 import com.asadabdalla.soccerplayer.models.Player;
 import com.asadabdalla.soccerplayer.repositories.PlayerRepository;
 import lombok.AllArgsConstructor;
@@ -18,10 +20,21 @@ public class PlayerService {
     }
 
     public void addPlayer(Player player) {
+        Boolean nameExist = playerRepository.selectNameExist(player.getName());
+        if(nameExist) {
+            throw new BadRequestException(
+                    "Player with name " + player.getName() + " already exist"
+            );
+        }
         playerRepository.save(player);
     }
 
     public void deletePlayer(Long playerId) {
+        if(!playerRepository.existsById(playerId)){
+            throw new PlayerNotFoundException(
+                    "Player with id " + playerId + " does not exist"
+            );
+        }
         playerRepository.deleteById(playerId);
     }
 }
